requirements.txt
beautifulsoup4==4.4.1
cycler==0.10.0
decorator==4.0.9
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
Pillow==3.2.0
PyMarvel==0.1.0
pyparsing==2.1.1
python-dateutil==2.5.2
pytz==2016.3
requests==2.9.1
six==1.10.0
wikia==1.4.3